+++ scripts/lib/utilities/tokenUtils.js
@@ -143,7 +143,7 @@ function getLightLevel(token) {
     if (!lights.length) return 'dark';
     let inBright = lights.some(light => {
         let {data: {x, y}, ratio} = light;
-        let bright = foundry.canvas.geometry.ClockwiseSweepPolygon.create({'x': x, 'y': y}, {
+        let bright = foundry.canvas.geometry.ClockwiseSweepPolygon.create({x: x, y: y}, {
             type: 'light',
             boundaryShapes: [new PIXI.Circle(x, y, ratio * light.shape.config.radius)]
         });
@@ -325,7 +325,7 @@ async function grappleHelper(sourceToken, targetToken, item, {noContest = false,
         timePassed += 100;
     }
     if (grappledEffect) await effectUtils.addDependent(grappledEffect, [targetEffect]);
-    if (game.modules.get('Rideable')?.active) game.Rideable.Mount([targetToken.document], sourceToken.document, {'Grappled': true, 'MountingEffectsOverride': ['Grappled']});
+    if (game.modules.get('Rideable')?.active) game.Rideable.Mount([targetToken.document], sourceToken.document, {Grappled: true, MountingEffectsOverride: ['Grappled']});
 }
 function isGrappledBy(target, source) {
     let effects = effectUtils.getAllEffectsByIdentifier(target.actor, 'grappled');