+++ scripts/macros/2014/raceFeatures/githzerai/mentalDiscipline.js
@@ -4,16 +4,16 @@ export let mentalDisciplineGithzerai = {
     ddbi: {
         restrictedItems: {
             'Mental Discipline 2': {
-                'originalName': 'Mental Discipline',
-                'requiredClass': null,
-                'requiredSubclass': null,
-                'requiredRace': 'Githzerai',
-                'requiredEquipment': [],
-                'requiredFeatures': [],
-                'replacedItemName': 'Mental Discipline (Githzerai)',
-                'removedItems': [],
-                'additionalItems': [],
-                'priority': 0
+                originalName: 'Mental Discipline',
+                requiredClass: null,
+                requiredSubclass: null,
+                requiredRace: 'Githzerai',
+                requiredEquipment: [],
+                requiredFeatures: [],
+                replacedItemName: 'Mental Discipline (Githzerai)',
+                removedItems: [],
+                additionalItems: [],
+                priority: 0
             }
         }
     }