+++ scripts/macros/2014/raceFeatures/kalashtar/mentalDiscipline.js
@@ -4,16 +4,16 @@ export let mentalDisciplineKalashtar = {
     ddbi: {
         restrictedItems: {
             'Mental Discipline 1': {
-                'originalName': 'Mental Discipline',
-                'requiredClass': null,
-                'requiredSubclass': null,
-                'requiredRace': 'Kalashtar',
-                'requiredEquipment': [],
-                'requiredFeatures': [],
-                'replacedItemName': 'Mental Discipline (Kalashtar)',
-                'removedItems': [],
-                'additionalItems': [],
-                'priority': 0
+                originalName: 'Mental Discipline',
+                requiredClass: null,
+                requiredSubclass: null,
+                requiredRace: 'Kalashtar',
+                requiredEquipment: [],
+                requiredFeatures: [],
+                replacedItemName: 'Mental Discipline (Kalashtar)',
+                removedItems: [],
+                additionalItems: [],
+                priority: 0
             },
         }
     }