+++ scripts/macros/2014/spells/disintegrate.js
@@ -14,9 +14,9 @@ async function use({workflow}) {
                 .atLocation(workflow.token)
                 .file('jb2a.magic_signs.circle.02.transmutation.loop.dark_green')
                 .scaleToObject(1.25)
-                .rotateIn(180, 600, {'ease': 'easeOutCubic'})
-                .scaleIn(0, 600, {'ease': 'easeOutCubic'})
-                .loopProperty('sprite', 'rotation', {'from': 0, 'to': -360, 'duration': 10000})
+                .rotateIn(180, 600, {ease: 'easeOutCubic'})
+                .scaleIn(0, 600, {ease: 'easeOutCubic'})
+                .loopProperty('sprite', 'rotation', {from: 0, to: -360, duration: 10000})
                 .belowTokens()
                 .fadeOut(2000)
                 .zIndex(0)
@@ -24,36 +24,36 @@ async function use({workflow}) {
                 .atLocation(workflow.token)
                 .file('jb2a.magic_signs.circle.02.transmutation.loop.dark_green')
                 .scaleToObject(1.25)
-                .rotateIn(180, 600, {'ease': 'easeOutCubic'})
-                .scaleIn(0, 600, {'ease': 'easeOutCubic'})
-                .loopProperty('sprite', 'rotation', {'from': 0, 'to': -360, 'duration': 10000})
+                .rotateIn(180, 600, {ease: 'easeOutCubic'})
+                .scaleIn(0, 600, {ease: 'easeOutCubic'})
+                .loopProperty('sprite', 'rotation', {from: 0, to: -360, duration: 10000})
                 .belowTokens(true)
-                .filter('ColorMatrix', {'saturate': -1, 'brightness': 2})
-                .filter('Blur', {'blurX': 5, 'blurY': 10 })
+                .filter('ColorMatrix', {saturate: -1, brightness: 2})
+                .filter('Blur', {blurX: 5, blurY: 10 })
                 .zIndex(1)
                 .duration(1200)
-                .fadeIn(200, {'ease': 'easeOutCirc', 'delay': 500})
-                .fadeOut(300, {'ease': 'linear'})
+                .fadeIn(200, {ease: 'easeOutCirc', delay: 500})
+                .fadeOut(300, {ease: 'linear'})
                 .zIndex(0.1)
             .effect()
                 .file('jb2a.particles.outward.white.01.02')
-                .scaleIn(0, 1000, {'ease': 'easeOutQuint'})
+                .scaleIn(0, 1000, {ease: 'easeOutQuint'})
                 .delay(500)
                 .fadeOut(1000)
                 .atLocation(workflow.token)
                 .duration(1000)
-                .size(1.75, {'gridUnits': true})
-                .animateProperty('spriteContainer', 'position.y', {'from':0 , 'to': -0.5, 'gridUnits': true, 'duration': 1000})
+                .size(1.75, {gridUnits: true})
+                .animateProperty('spriteContainer', 'position.y', {from:0 , to: -0.5, gridUnits: true, duration: 1000})
                 .zIndex(1)
             .effect()
                 .file('jb2a.particles.outward.white.01.02')
-                .scaleIn(0, 1000, {'ease': 'easeOutQuint'})
+                .scaleIn(0, 1000, {ease: 'easeOutQuint'})
                 .delay(500)
                 .fadeOut(1000)
                 .atLocation(workflow.token)
                 .duration(1000)
-                .size(1.75, {'gridUnits': true})
-                .animateProperty('spriteContainer', 'position.y', {'from': 0 , 'to': -0.5, 'gridUnits': true, 'duration': 1000})
+                .size(1.75, {gridUnits: true})
+                .animateProperty('spriteContainer', 'position.y', {from: 0 , to: -0.5, gridUnits: true, duration: 1000})
                 .zIndex(1)
                 .mirrorX()
             .wait(1000)
@@ -81,14 +81,14 @@ async function use({workflow}) {
                 .playIf(targetDeath)
             .effect()
                 .file('animated-spell-effects-cartoon.smoke.97')
-                .atLocation(target, {'offset': {'y':-0.25}, 'gridUnits': true})
+                .atLocation(target, {offset: {y:-0.25}, gridUnits: true})
                 .fadeIn(1000)
-                .scaleIn(0, 1000, {'ease': 'easeOutCubic'})
+                .scaleIn(0, 1000, {ease: 'easeOutCubic'})
                 .delay(1000)
                 .duration(10000)
                 .fadeOut(500)
                 .scaleToObject(0.5)
-                .filter('ColorMatrix', {'brightness': 0})
+                .filter('ColorMatrix', {brightness: 0})
                 .zIndex(0.1)
                 .belowTokens()
                 .playIf(targetDeath)
@@ -98,7 +98,7 @@ async function use({workflow}) {
                 .duration(7500)
                 .fadeOut(3000)
                 .scaleToObject(0.35)
-                .filter('ColorMatrix', {'hue': -25})
+                .filter('ColorMatrix', {hue: -25})
                 .belowTokens()
                 .playIf(targetDeath)
             .effect()
@@ -106,14 +106,14 @@ async function use({workflow}) {
                 .copySprite(target)
                 .scaleToObject(target.document.texture.scaleX)
                 .shape('circle', {
-                    'lineSize': 25,
-                    'lineColor': '#FF0000',
-                    'fillColor': '#FF0000',
-                    'radius': 0.15,
-                    'gridUnits': true,
-                    'name': 'test',
-                    'isMask': true,
-                    'offset': {'x': canvas.grid.size * 0.1, 'y': -canvas.grid.size * 0.4}, 
+                    lineSize: 25,
+                    lineColor: '#FF0000',
+                    fillColor: '#FF0000',
+                    radius: 0.15,
+                    gridUnits: true,
+                    name: 'test',
+                    isMask: true,
+                    offset: {x: canvas.grid.size * 0.1, y: -canvas.grid.size * 0.4}, 
                     
                 })
                 .duration(1500)
@@ -124,13 +124,13 @@ async function use({workflow}) {
                 .copySprite(target)
                 .scaleToObject(target.document.texture.scaleX)
                 .shape('circle', {
-                    'lineSize': 25,
-                    'lineColor': '#FF0000',
-                    'radius': 0.2,
-                    'gridUnits': true,
-                    'name': 'test',
-                    'isMask': true,
-                    'offset': {'x': canvas.grid.size * 0.1, 'y': -canvas.grid.size * 0.4},            
+                    lineSize: 25,
+                    lineColor: '#FF0000',
+                    radius: 0.2,
+                    gridUnits: true,
+                    name: 'test',
+                    isMask: true,
+                    offset: {x: canvas.grid.size * 0.1, y: -canvas.grid.size * 0.4},            
                 })
                 .duration(1800)
                 .fadeOut(1000)
@@ -140,13 +140,13 @@ async function use({workflow}) {
                 .copySprite(target)
                 .scaleToObject(target.document.texture.scaleX)
                 .shape('circle', {
-                    'lineSize': 25,
-                    'lineColor': '#FF0000',
-                    'radius': 0.25,
-                    'gridUnits': true,
-                    'name': 'test',
-                    'isMask': true,
-                    'offset': {'x': canvas.grid.size * 0.1, 'y': -canvas.grid.size * 0.4}, 
+                    lineSize: 25,
+                    lineColor: '#FF0000',
+                    radius: 0.25,
+                    gridUnits: true,
+                    name: 'test',
+                    isMask: true,
+                    offset: {x: canvas.grid.size * 0.1, y: -canvas.grid.size * 0.4}, 
                 })
                 .duration(2000)
                 .fadeOut(1000)
@@ -156,13 +156,13 @@ async function use({workflow}) {
                 .copySprite(target)
                 .scaleToObject(target.document.texture.scaleX)
                 .shape('circle', {
-                    'lineSize': 25,
-                    'lineColor': '#FF0000',
-                    'radius': 0.3,
-                    'gridUnits': true,
-                    'name': 'test',
-                    'isMask': true,
-                    'offset': {'x': canvas.grid.size * 0.1, 'y': -canvas.grid.size * 0.4}, 
+                    lineSize: 25,
+                    lineColor: '#FF0000',
+                    radius: 0.3,
+                    gridUnits: true,
+                    name: 'test',
+                    isMask: true,
+                    offset: {x: canvas.grid.size * 0.1, y: -canvas.grid.size * 0.4}, 
                 })
                 .duration(2200)
                 .fadeOut(1000)
@@ -172,13 +172,13 @@ async function use({workflow}) {
                 .copySprite(target)
                 .scaleToObject(target.document.texture.scaleX)
                 .shape('circle', {
-                    'lineSize': 25,
-                    'lineColor': '#FF0000',
-                    'radius': 0.35,
-                    'gridUnits': true,
-                    'name': 'test',
-                    'isMask': true,
-                    'offset': {'x': canvas.grid.size * 0.1, 'y': -canvas.grid.size * 0.4}, 
+                    lineSize: 25,
+                    lineColor: '#FF0000',
+                    radius: 0.35,
+                    gridUnits: true,
+                    name: 'test',
+                    isMask: true,
+                    offset: {x: canvas.grid.size * 0.1, y: -canvas.grid.size * 0.4}, 
                 })   
                 .duration(2400)
                 .fadeOut(1000)
@@ -188,13 +188,13 @@ async function use({workflow}) {
                 .copySprite(target)
                 .scaleToObject(target.document.texture.scaleX)
                 .shape('circle', {
-                    'lineSize': 25,
-                    'lineColor': '#FF0000',
-                    'radius': 0.4,
-                    'gridUnits': true,
-                    'name': 'test',
-                    'isMask': true,
-                    'offset': {'x': canvas.grid.size * 0.1, 'y': -canvas.grid.size * 0.4}, 
+                    lineSize: 25,
+                    lineColor: '#FF0000',
+                    radius: 0.4,
+                    gridUnits: true,
+                    name: 'test',
+                    isMask: true,
+                    offset: {x: canvas.grid.size * 0.1, y: -canvas.grid.size * 0.4}, 
                 })
                 .duration(2600)
                 .fadeOut(1000)
@@ -204,13 +204,13 @@ async function use({workflow}) {
                 .copySprite(target)
                 .scaleToObject(target.document.texture.scaleX)
                 .shape('circle', {
-                    'lineSize': 25,
-                    'lineColor': '#FF0000',
-                    'radius': 0.45,
-                    'gridUnits': true,
-                    'name': 'test',
-                    'isMask': true,
-                    'offset': {'x': canvas.grid.size * 0.1, 'y': -canvas.grid.size * 0.4}, 
+                    lineSize: 25,
+                    lineColor: '#FF0000',
+                    radius: 0.45,
+                    gridUnits: true,
+                    name: 'test',
+                    isMask: true,
+                    offset: {x: canvas.grid.size * 0.1, y: -canvas.grid.size * 0.4}, 
                 })    
                 .duration(2800)
                 .fadeOut(1000)
@@ -220,14 +220,14 @@ async function use({workflow}) {
                 .copySprite(target)
                 .scaleToObject(target.document.texture.scaleX)
                 .shape('circle', {
-                    'lineSize': 25,
-                    'lineColor': '#FF0000',
-                    'fillColor': '#FF0000',
-                    'radius': 0.15,
-                    'gridUnits': true,
-                    'name': 'test',
-                    'isMask': true,
-                    'offset': {'x': -canvas.grid.size * 0.4, 'y': canvas.grid.size * 0.3}, 
+                    lineSize: 25,
+                    lineColor: '#FF0000',
+                    fillColor: '#FF0000',
+                    radius: 0.15,
+                    gridUnits: true,
+                    name: 'test',
+                    isMask: true,
+                    offset: {x: -canvas.grid.size * 0.4, y: canvas.grid.size * 0.3}, 
                     
                 })
                 .duration(500)
@@ -238,13 +238,13 @@ async function use({workflow}) {
                 .copySprite(target)
                 .scaleToObject(target.document.texture.scaleX)
                 .shape('circle', {
-                    'lineSize': 25,
-                    'lineColor': '#FF0000',
-                    'radius': 0.2,
-                    'gridUnits': true,
-                    'name': 'test',
-                    'isMask': true,
-                    'offset': {'x': -canvas.grid.size * 0.4, 'y': canvas.grid.size * 0.3},           
+                    lineSize: 25,
+                    lineColor: '#FF0000',
+                    radius: 0.2,
+                    gridUnits: true,
+                    name: 'test',
+                    isMask: true,
+                    offset: {x: -canvas.grid.size * 0.4, y: canvas.grid.size * 0.3},           
                 })
                 .duration(700)
                 .fadeOut(1000)
@@ -254,13 +254,13 @@ async function use({workflow}) {
                 .copySprite(target)
                 .scaleToObject(target.document.texture.scaleX)
                 .shape('circle', {
-                    'lineSize': 25,
-                    'lineColor': '#FF0000',
-                    'radius': 0.25,
-                    'gridUnits': true,
-                    'name': 'test',
-                    'isMask': true,
-                    'offset': {'x': -canvas.grid.size * 0.4, 'y': canvas.grid.size * 0.3}, 
+                    lineSize: 25,
+                    lineColor: '#FF0000',
+                    radius: 0.25,
+                    gridUnits: true,
+                    name: 'test',
+                    isMask: true,
+                    offset: {x: -canvas.grid.size * 0.4, y: canvas.grid.size * 0.3}, 
                 })
                 .duration(900)
                 .fadeOut(1000)
@@ -270,13 +270,13 @@ async function use({workflow}) {
                 .copySprite(target)
                 .scaleToObject(target.document.texture.scaleX)
                 .shape('circle', {
-                    'lineSize': 25,
-                    'lineColor': '#FF0000',
-                    'radius': 0.3,
-                    'gridUnits': true,
-                    'name': 'test',
-                    'isMask': true,
-                    'offset': {'x': -canvas.grid.size * 0.4, 'y': canvas.grid.size * 0.3}, 
+                    lineSize: 25,
+                    lineColor: '#FF0000',
+                    radius: 0.3,
+                    gridUnits: true,
+                    name: 'test',
+                    isMask: true,
+                    offset: {x: -canvas.grid.size * 0.4, y: canvas.grid.size * 0.3}, 
                 })
                 .duration(1100)
                 .fadeOut(1000)
@@ -286,13 +286,13 @@ async function use({workflow}) {
                 .copySprite(target)
                 .scaleToObject(target.document.texture.scaleX)
                 .shape('circle', {
-                    'lineSize': 25,
-                    'lineColor': '#FF0000',
-                    'radius': 0.35,
-                    'gridUnits': true,
-                    'name': 'test',
-                    'isMask': true,
-                    'offset': {'x': -canvas.grid.size * 0.4, 'y': canvas.grid.size * 0.3}, 
+                    lineSize: 25,
+                    lineColor: '#FF0000',
+                    radius: 0.35,
+                    gridUnits: true,
+                    name: 'test',
+                    isMask: true,
+                    offset: {x: -canvas.grid.size * 0.4, y: canvas.grid.size * 0.3}, 
                 })   
                 .duration(1300)
                 .fadeOut(1000)
@@ -302,13 +302,13 @@ async function use({workflow}) {
                 .copySprite(target)
                 .scaleToObject(target.document.texture.scaleX)
                 .shape('circle', {
-                    'lineSize': 25,
-                    'lineColor': '#FF0000',
-                    'radius': 0.4,
-                    'gridUnits': true,
-                    'name': 'test',
-                    'isMask': true,
-                    'offset': {'x': -canvas.grid.size * 0.4, 'y': canvas.grid.size * 0.3}, 
+                    lineSize: 25,
+                    lineColor: '#FF0000',
+                    radius: 0.4,
+                    gridUnits: true,
+                    name: 'test',
+                    isMask: true,
+                    offset: {x: -canvas.grid.size * 0.4, y: canvas.grid.size * 0.3}, 
                 })
                 .duration(1500)
                 .fadeOut(1000)
@@ -318,13 +318,13 @@ async function use({workflow}) {
                 .copySprite(target)
                 .scaleToObject(target.document.texture.scaleX)
                 .shape('circle', {
-                    'lineSize': 25,
-                    'lineColor': '#FF0000',
-                    'radius': 0.45,
-                    'gridUnits': true,
-                    'name': 'test',
-                    'isMask': true,
-                    'offset': {'x': -canvas.grid.size * 0.4, 'y': canvas.grid.size * 0.3}, 
+                    lineSize: 25,
+                    lineColor: '#FF0000',
+                    radius: 0.45,
+                    gridUnits: true,
+                    name: 'test',
+                    isMask: true,
+                    offset: {x: -canvas.grid.size * 0.4, y: canvas.grid.size * 0.3}, 
                 })   
                 .duration(1700)
                 .fadeOut(1000)
@@ -334,13 +334,13 @@ async function use({workflow}) {
                 .copySprite(target)
                 .scaleToObject(target.document.texture.scaleX)
                 .shape('circle', {
-                    'lineSize': 25,
-                    'lineColor': '#FF0000',
-                    'radius': 0.5,
-                    'gridUnits': true,
-                    'name': 'test',
-                    'isMask': true,
-                    'offset': {'x': -canvas.grid.size * 0.4, 'y': canvas.grid.size * 0.3}, 
+                    lineSize: 25,
+                    lineColor: '#FF0000',
+                    radius: 0.5,
+                    gridUnits: true,
+                    name: 'test',
+                    isMask: true,
+                    offset: {x: -canvas.grid.size * 0.4, y: canvas.grid.size * 0.3}, 
                 }) 
                 .duration(1900)
                 .fadeOut(1000)
@@ -350,13 +350,13 @@ async function use({workflow}) {
                 .copySprite(target)
                 .scaleToObject(target.document.texture.scaleX)
                 .shape('circle', {
-                    'lineSize': 25,
-                    'lineColor': '#FF0000',
-                    'radius': 0.55,
-                    'gridUnits': true,
-                    'name': 'test',
-                    'isMask': true,
-                    'offset': {'x': -canvas.grid.size * 0.4, 'y' :canvas.grid.size * 0.3}, 
+                    lineSize: 25,
+                    lineColor: '#FF0000',
+                    radius: 0.55,
+                    gridUnits: true,
+                    name: 'test',
+                    isMask: true,
+                    offset: {x: -canvas.grid.size * 0.4, y :canvas.grid.size * 0.3}, 
                 })
                 .duration(2100)
                 .fadeOut(1000)
@@ -366,14 +366,14 @@ async function use({workflow}) {
                 .copySprite(target)
                 .scaleToObject(target.document.texture.scaleX)
                 .shape('circle', {
-                    'lineSize': 25,
-                    'lineColor': '#FF0000',
-                    'fillColor': '#FF0000',
-                    'radius': 0.15,
-                    'gridUnits': true,
-                    'name': 'test',
-                    'isMask': true,
-                    'offset': {'x': canvas.grid.size*0.5, 'y': canvas.grid.size*0.4}, 
+                    lineSize: 25,
+                    lineColor: '#FF0000',
+                    fillColor: '#FF0000',
+                    radius: 0.15,
+                    gridUnits: true,
+                    name: 'test',
+                    isMask: true,
+                    offset: {x: canvas.grid.size*0.5, y: canvas.grid.size*0.4}, 
                 })
                 .duration(1500)
                 .fadeOut(1000)
@@ -383,13 +383,13 @@ async function use({workflow}) {
                 .copySprite(target)
                 .scaleToObject(target.document.texture.scaleX) 
                 .shape('circle', {
-                    'lineSize': 25,
-                    'lineColor': '#FF0000',
-                    'radius': 0.2,
-                    'gridUnits': true,
-                    'name': 'test',
-                    'isMask': true,
-                    'offset': {'x': canvas.grid.size * 0.5, 'y': canvas.grid.size * 0.4},  
+                    lineSize: 25,
+                    lineColor: '#FF0000',
+                    radius: 0.2,
+                    gridUnits: true,
+                    name: 'test',
+                    isMask: true,
+                    offset: {x: canvas.grid.size * 0.5, y: canvas.grid.size * 0.4},  
                 })
                 .duration(1700)
                 .fadeOut(1000)
@@ -399,13 +399,13 @@ async function use({workflow}) {
                 .copySprite(target)
                 .scaleToObject(target.document.texture.scaleX) 
                 .shape('circle', {
-                    'lineSize': 25,
-                    'lineColor': '#FF0000',
-                    'radius': 0.25,
-                    'gridUnits': true,
-                    'name': 'test',
-                    'isMask': true,
-                    'offset': {'x': canvas.grid.size * 0.5, 'y': canvas.grid.size * 0.4},  
+                    lineSize: 25,
+                    lineColor: '#FF0000',
+                    radius: 0.25,
+                    gridUnits: true,
+                    name: 'test',
+                    isMask: true,
+                    offset: {x: canvas.grid.size * 0.5, y: canvas.grid.size * 0.4},  
                 })
                 .duration(1900)
                 .fadeOut(1000)
@@ -415,13 +415,13 @@ async function use({workflow}) {
                 .copySprite(target)
                 .scaleToObject(target.document.texture.scaleX)
                 .shape('circle', {
-                    'lineSize': 25,
-                    'lineColor': '#FF0000',
-                    'radius': 0.3,
-                    'gridUnits': true,
-                    'name': 'test',
-                    'isMask': true,
-                    'offset': {'x': canvas.grid.size * 0.5, 'y': canvas.grid.size * 0.4}, 
+                    lineSize: 25,
+                    lineColor: '#FF0000',
+                    radius: 0.3,
+                    gridUnits: true,
+                    name: 'test',
+                    isMask: true,
+                    offset: {x: canvas.grid.size * 0.5, y: canvas.grid.size * 0.4}, 
                 })
                 .duration(2100)
                 .fadeOut(1000)
@@ -431,13 +431,13 @@ async function use({workflow}) {
                 .copySprite(target)
                 .scaleToObject(target.document.texture.scaleX)
                 .shape('circle', {
-                    'lineSize': 25,
-                    'lineColor': '#FF0000',
-                    'radius': 0.35,
-                    'gridUnits': true,
-                    'name': 'test',
-                    'isMask': true,
-                    'offset': {'x': canvas.grid.size * 0.5, 'y': canvas.grid.size * 0.4}, 
+                    lineSize: 25,
+                    lineColor: '#FF0000',
+                    radius: 0.35,
+                    gridUnits: true,
+                    name: 'test',
+                    isMask: true,
+                    offset: {x: canvas.grid.size * 0.5, y: canvas.grid.size * 0.4}, 
                 })   
                 .duration(2300)
                 .fadeOut(1000)
@@ -447,13 +447,13 @@ async function use({workflow}) {
                 .copySprite(target)
                 .scaleToObject(target.document.texture.scaleX)
                 .shape('circle', {
-                    'lineSize': 25,
-                    'lineColor': '#FF0000',
-                    'radius': 0.4,
-                    'gridUnits': true,
-                    'name': 'test',
-                    'isMask': true,
-                    'offset': {'x': canvas.grid.size * 0.5, 'y': canvas.grid.size * 0.4}, 
+                    lineSize: 25,
+                    lineColor: '#FF0000',
+                    radius: 0.4,
+                    gridUnits: true,
+                    name: 'test',
+                    isMask: true,
+                    offset: {x: canvas.grid.size * 0.5, y: canvas.grid.size * 0.4}, 
                 })
                 .duration(2500)
                 .fadeOut(1000)
@@ -462,13 +462,13 @@ async function use({workflow}) {
                 .copySprite(target)
                 .scaleToObject(target.document.texture.scaleX)
                 .shape('circle', {
-                    'lineSize': 25,
-                    'lineColor': '#FF0000',
-                    'radius': 0.45,
-                    'gridUnits': true,
-                    'name': 'test',
-                    'isMask': true,
-                    'offset': {'x': canvas.grid.size * 0.5, 'y': canvas.grid.size * 0.4}, 
+                    lineSize: 25,
+                    lineColor: '#FF0000',
+                    radius: 0.45,
+                    gridUnits: true,
+                    name: 'test',
+                    isMask: true,
+                    offset: {x: canvas.grid.size * 0.5, y: canvas.grid.size * 0.4}, 
                 })    
                 .duration(2700)
                 .fadeOut(1000)