+++ scripts/macros/2014/spells/healingSpirit.js
@@ -168,7 +168,7 @@ async function use({workflow}) {
                 .atLocation(spawnedToken)
                 .scaleToObject(1.5)
                 .startTime(1500)
-                .filter('ColorMatrix', {'hue': 15})
+                .filter('ColorMatrix', {hue: 15})
             .play();
             /* eslint-enable indent */
     }