+++ scripts/macros/2024/actions/help.js
@@ -1,6 +1,6 @@
 import {help as legacyHelp} from '../../2014/actions/help.js';
 export let help = {
     ...legacyHelp,
-    version: '1.3.115',
+    version: '1.3.139',
     rules: 'modern'
 };