+++ scripts/macros/2024/classFeatures/monk/uncannyMetabolism.js
@@ -1,0 +1,86 @@
+import {itemUtils} from '../../../../utils.js';
+async function added({trigger: {entity: item}}) {
+    await itemUtils.correctActivityItemConsumption(item, ['heal'], 'martialArts');
+    await itemUtils.fixScales(item);
+}
+export let uncannyMetabolism = {
+    name: 'Uncanny Metabolism',
+    version: '1.3.138',
+    rules: 'modern',
+    item: [
+        {
+            pass: 'created',
+            macro: added,
+            priority: 45
+        },
+        {
+            pass: 'itemMedkit',
+            macro: added,
+            priority: 45
+        },
+        {
+            pass: 'actorMunch',
+            macro: added,
+            priority: 45
+        }
+    ],
+    config: [
+        {
+            value: 'classIdentifier',
+            label: 'CHRISPREMADES.Config.ClassIdentifier',
+            type: 'text',
+            default: 'monk',
+            category: 'homebrew',
+            homebrew: true
+        },
+        {
+            value: 'scaleIdentifier',
+            label: 'CHRISPREMADES.Config.ScaleIdentifier',
+            type: 'text',
+            default: 'martial-arts',
+            category: 'homebrew',
+            homebrew: true
+        }
+    ],
+    scales: [
+        {
+            classIdentifier: 'classIdentifier',
+            scaleIdentifier: 'scaleIdentifier',
+            data: {
+                type: 'ScaleValue',
+                configuration: {
+                    identifier: 'martial-arts',
+                    type: 'dice',
+                    distance: {
+                        units: ''
+                    },
+                    scale: {
+                        1: {
+                            number: 1,
+                            faces: 6,
+                            modifiers: []
+                        },
+                        5: {
+                            number: 1,
+                            faces: 8,
+                            modifiers: []
+                        },
+                        11: {
+                            number: 1,
+                            faces: 10,
+                            modifiers: []
+                        },
+                        17: {
+                            number: 1,
+                            faces: 12,
+                            modifiers: []
+                        }
+                    }
+                },
+                value: {},
+                title: 'Martial Arts'
+            }
+        }
+        
+    ]
+};