+++ scripts/macros/2024/spells/protectionFromEvilAndGood.js
@@ -1,6 +1,6 @@
 import {protectionFromEvilAndGood as protectionFromEvilAndGoodLegacy} from '../../../legacyMacros.js';
 export let protectionFromEvilAndGood = {
     ...protectionFromEvilAndGoodLegacy,
-    version: '1.3.129',
+    version: '1.3.138',
     rules: 'modern'
 };